refactor(isolate): 지갑용과 서명용 isolate class 생성.
isolate용 함수는 상태가 없는 클래스의 독립 함수여야 함.
