+++ lib/model/multisig/multisig_vault_list_item.dart
@@ -1,10 +1,10 @@
 import 'dart:convert';
 
 import 'package:coconut_lib/coconut_lib.dart';
+import 'package:coconut_vault/isolates/sign_isolates.dart';
 import 'package:coconut_vault/model/multisig/multisig_signer.dart';
 import 'package:coconut_vault/model/common/vault_list_item_base.dart';
 import 'package:coconut_vault/enums/wallet_enums.dart';
-import 'package:coconut_vault/managers/isolate_manager.dart';
 import 'package:coconut_vault/utils/isolate_handler.dart';
 import 'package:json_annotation/json_annotation.dart';
 
@@ -44,7 +44,7 @@ class MultisigVaultListItem extends VaultListItemBase {
 
   @override
   Future<bool> canSign(String psbt) async {
-    var isolateHandler = IsolateHandler<List<dynamic>, bool>(canSignToPsbtIsolate);
+    var isolateHandler = IsolateHandler<List<dynamic>, bool>(SignIsolates.canSignToPsbt);
     try {
       await isolateHandler.initialize(initialType: InitializeType.canSign);
       bool canSignToPsbt = await isolateHandler.run([coconutVault, psbt]);
